+++ src/environment.c
@@ -431,7 +431,7 @@ env_t *global_env(bool source_mapping, bool profiling, bool debugging) {
             {"by_split_any", "Text$by_split_any", "func(text:Text, delimiters=' \\t\\r\\n' -> func(->Text?))"}, //
             {"caseless_equals", "Text$equal_ignoring_case", "func(a,b:Text, language='C' -> Bool)"}, //
             {"codepoint_names", "Text$codepoint_names", "func(text:Text -> [Text])"}, //
-            {"distance", "Text$distance", "func(a,b:Text, language='C' -> Num)"}, //
+            {"distance", "Text$distance", "func(a,b:Text -> Num)"}, //
             {"nearest", "Text$nearest", "func(text:Text, candidates:[Text], max_distance=0.6 -> Text?)"}, //
             {"ends_with", "Text$ends_with", "func(text,suffix:Text, remainder:&Text? = none -> Bool)"}, //
             {"find", "Text$find", "func(text,target:Text, start=1 -> Int?)"}, //